Reduce stack instantiation parameters
This commit reduces the parameters needed to instantiate a Stack. Stacks
are now instantiated with StackConfigData object. A future change will
introduce a the concept of a StackSchema so that we can define Stacks
from different providers, at that stage Stacks will be instantiated with
a StackConfigData and a Schema.

StackConfigData is pretty basic at the moment and just accepts kwargs
and adds these to the object __dict__. This will very likely need
improved before making it into production but for now allows us to move
forward.

This commit also removed `get_external_name` from helpers. This removes
the dependency between `helpers` module and `providers` module and can
be implemented if required by individual provider packages.