Move pytest addopts from tox.ini to Makefile
Moved the `addopts` pytest config from tox.ini to Makefile as postargs
since having addopts in tox.ini can cause a conflict with pytest and
running individual tests. The previous config was only used as part of
the CI/CD pipeline and the new change will not affect the output or
functionality of our pipeline.