Add Command and Provider Command Regsitry
This commit enables users to create commands that automatically register
themselves against a Providers "Command Registry".

Any command a users creates should implement `Command` and the class be
follow the naming style VerbNoun. For example, if a user wished to
implement a command that created a new Stack in the AwsProvider the
command would be as follows:

```
class CreateStack(Command, provider=AwsProvider)
   def execute(cls):
      ... implementation here ...
```

This command class would then be available at:
AwsProvider.command_registry['create'][create_stack]

The command registry is first keyed by the verb defined in the
class name and then has additional key of the full class name.

e.g.
```
{
 "create": {
    "create_stack": CreateStack,
    "create_changeset": CreateChangeSet
 }
 "update": {
    "update_stack": UpdateStack,
    "update_changeset": UpdateChangeSet
 }
}
```

The benefit of this approach is that Providers can define their own
commands and there is only one reason for each command class to change.
It also eliminates the need to have a well defined interface of commands
at the core level.

In addition Sceptre also has a reference to every available command from
each of the registered providers in advance of a command being
triggered. This enables more useful error messages and assistance on the
CLI.